Move builtin and gazetteer parsers into dedicated files
