+++ platforms/snips-nlu-ontology-python/snips_nlu_ontology/__init__.py
@@ -1,8 +1,9 @@
 from __future__ import absolute_import
 
 from snips_nlu_ontology.builtin_entities import (
-    BuiltinEntityParser, get_all_builtin_entities, get_all_gazetteer_entities,
+    get_all_builtin_entities, get_all_gazetteer_entities,
     get_all_grammar_entities, get_all_languages, get_builtin_entity_examples,
     get_builtin_entity_shortname, get_ontology_version, get_supported_entities,
     get_supported_gazetteer_entities, get_supported_grammar_entities)
-from snips_nlu_ontology.gazetteer_entities import GazetteerEntityParser
+from snips_nlu_ontology.builtin_entity_parser import BuiltinEntityParser
+from snips_nlu_ontology.gazetteer_entity_parser import GazetteerEntityParser
