+++ platforms/snips-nlu-ontology-python/snips_nlu_ontology/gazetteer_entity_parser.py
